moved `get_pixel_colours` to omp/base.py for clarity
